refactoring metacritic-webscraper
